32-bit compile fix

+++ renderdoc/driver/d3d12/d3d12_device_wrap14.cpp
@@ -125,7 +125,7 @@ bool WrappedID3D12Device::Serialise_CreateRootSignatureFromSubobjectInLibrary(
 
       WrappedID3D12RootSignature *wrapped = (WrappedID3D12RootSignature *)ret;
 
-      wrapped->sig = DecodeRootSig(pLibraryBlob, blobLengthInBytes, subobjectName);
+      wrapped->sig = DecodeRootSig(pLibraryBlob, (size_t)blobLengthInBytes, subobjectName);
 
       if(wrapped->sig.Flags & D3D12_ROOT_SIGNATURE_FLAG_LOCAL_ROOT_SIGNATURE)
         wrapped->localRootSigIdx =
